bitmap::text

+++ library/base.h
@@ -12,5 +12,5 @@
 using uchar = unsigned char;
 using ushort = unsigned short;
 using uint = unsigned int;
-using u64 = unsigned long long;
-using i64 = long long;
+using u64 = uint64_t;
+using i64 = int64_t;
